refs #21 - Made project and project step forms use bootstrap and placeholders like the signup form
