+++ raspberryio/templates/project/project_step_create_edit.html
@@ -1,5 +1,6 @@
 {% extends "base.html" %}
 {% load url from future %}
+{% load bootstrap_toolkit %}
 
 {% block main %}
     <h3>
@@ -14,7 +15,7 @@
     {% include 'includes/file_upload/uploader_form.html' %}
 
     <form id="project-step-form" class="project-step-create-edit-form" action="" method="post" accept-charset="utf-8"> {% csrf_token %}
-        {{ project_step_form.as_p }}
+        {{ project_step_form|as_bootstrap }}
         <input type="submit" name="save" id="submit-project-step-create-edit" value="Save and View"/>
         <input type="submit" name="save-add" id="submit-project-step-create-edit-add-another" value="Save and add another" />
     </form>
